+++ src/lib.rs
@@ -1,9 +1,10 @@
 #![doc = include_str!("../README.md")]
-mod impls;
+mod maps;
+mod sets;
 mod try_collect_ex;
 mod try_extend;
 mod try_from_iterator;
 
 pub use try_collect_ex::TryCollectEx;
 pub use try_extend::TryExtend;
-pub use try_from_iterator::{KeyCollision, TryFromIterator};
+pub use try_from_iterator::{KeyCollision, TryFromIterator, ValueCollision};