+++ src/maps/mod.rs
@@ -1,8 +1,5 @@
-#[cfg(feature = "hash_map")]
-pub mod hash_map;
-
-#[cfg(feature = "btree_map")]
 pub mod btree_map;
+pub mod hash_map;
 
 #[cfg(feature = "hashbrown")]
 pub mod hashbrown;