+++ src/try_from_iterator.rs
@@ -33,3 +33,10 @@ pub struct KeyCollision<K> {
     /// The key that caused the collision.
     pub key: K,
 }
+
+/// An error indicating that a duplicate value was found in the provided data.
+#[derive(Debug, thiserror::Error, derive_more::Constructor, PartialEq, Eq, PartialOrd, Ord)]
+pub struct ValueCollision<T> {
+    /// The value that caused the collision.
+    pub value: T,
+}